Devicons: Keep more icons (and update dropbox)
[why]
Some icons that have been dropped from devicons are in wide use and so
we want to keep those at their old codepoints.

The Vorillaz dropbox logo is very old. As the icons we keep from
Vorillaz will not be maintained by any upstream we have the
responsibility to keep it up to date.

[how]
Put the icons in the appropriate folder; icon files taken out of
vorillaz/devicons `!SVG/` directory.

Adapt rules and mapping.

The Dropbox icon has been hand-modified (simplified) with the original
logo as starting point.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>